requirements.txt
tensorflow==2.12.0
matplotlib==3.7.1
keras==2.12
plotly
python-dotenv
pandas
dash
h5py==3.8
scikit-learn==1.2.2
scipy==1.10.1
dash_bootstrap_components
ipynb
torchsummary
functorch
chipwhisperer
pytest==6.2.5
pytest-pythonpath==0.7.4
pytest-cov==4.0.0
grad-cam==1.4.6
pycryptodome
python-dotenv
imbalanced-learn==0.10.1
seaborn==0.12.2
protobuf
sphinx
sphinxcontrib-bibtex
sphinx-rtd-theme
numpy